Move compress/decompress to Extensions.kt Makes more sense, since they can apply to data outside of App Manager
