+++ common/src/main/kotlin/com/boswelja/smartwatchextensions/common/Extensions.kt
@@ -1,5 +1,8 @@
 package com.boswelja.smartwatchextensions.common
 
+import java.util.zip.Deflater
+import java.util.zip.Inflater
+
 /**
  * Convert a single [Boolean] to a [ByteArray].
  * @return The created [ByteArray].
@@ -54,4 +57,30 @@ fun Int.Companion.fromByteArray(byteArray: ByteArray): Int {
             (0xff and byteArray[2].toInt() shl 8) or
             (0xff and byteArray[3].toInt() shl 0)
     }
+}
+
+fun ByteArray.compress(
+    level: Int = Deflater.BEST_COMPRESSION
+): ByteArray {
+    // Start compression
+    val compressor = Deflater(level)
+    compressor.setInput(this)
+    compressor.finish()
+
+    // Get compressed bytes
+    val result = ByteArray(size) // Assume we'll never exceed the size of the original bytes
+    val totalBytes = compressor.deflate(result)
+    compressor.end()
+
+    return result.copyOf(totalBytes)
+}
+
+fun ByteArray.decompress(): ByteArray {
+    // Decompress bytes
+    val decompressor = Inflater()
+    decompressor.setInput(this)
+    val decompressedBytes = ByteArray(Short.MAX_VALUE.toInt())
+    val resultLength = decompressor.inflate(decompressedBytes)
+    decompressor.end()
+    return decompressedBytes.copyOf(resultLength)
 }
